+++ examples/slider.rs
@@ -1,0 +1,32 @@
+extern crate pushrod;
+extern crate sdl2;
+
+use pushrod::render::engine::Engine;
+use pushrod::render::widget::Widget;
+use pushrod::render::widget_config::{CONFIG_BORDER_WIDTH, CONFIG_COLOR_BASE, CONFIG_COLOR_BORDER};
+use pushrod::widgets::toggle_button_widget::ToggleButtonWidget;
+use sdl2::pixels::Color;
+use pushrod::widgets::slider_widget::{SliderWidget, SliderOrientation};
+use pushrod::widgets::slider_widget::SliderOrientation::SliderHorizontal;
+
+/*
+ * This demo just tests the rendering functionality of the `BaseWidget`.  It only tests the
+ * render portion of the library, nothing else.
+ */
+
+pub fn main() {
+    let sdl_context = sdl2::init().unwrap();
+    let video_subsystem = sdl_context.video().unwrap();
+    let window = video_subsystem
+        .window("pushrod-render toggle button demo", 400, 100)
+        .position_centered()
+        .opengl()
+        .build()
+        .unwrap();
+    let mut engine = Engine::new(400, 100);
+    let mut slider1 = SliderWidget::new(20, 20, 360, 20, 0, 100, 20, SliderHorizontal);
+
+    engine.add_widget(Box::new(slider1), String::from("slider1"));
+
+    engine.run(sdl_context, window);
+}